requirements.txt
django>=3.2.0
djangorestframework>=3.14.0
gunicorn==23.0.0